requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
langchain==0.1.9
langchain-google-genai==0.0.11
tavily-python==0.3.9
reportlab==4.1.0
python-dotenv==1.0.1
streamlit==1.31.1